add image-input component, split main file-upload component
