+++ frontend/src/app/features/components/file-upload/image-input/image-input.component.css
