+++ frontend/src/app/features/components/file-upload/image-input/image-input.component.html
@@ -49,5 +49,3 @@
 </div>
 
 <div *ngIf="message" class="alert alert-secondary" role="alert">{{message}}</div>
-
-<app-image-list></app-image-list>
